+++ src/generation/flat/flatTerrainGenerator.cpp
@@ -8,7 +8,7 @@ Heightmap FlatTerrainGenerator::generate(Vector2<unsigned int> dimensions){
     Heightmap heightmap;   
     unsigned int height = 4; //standard random number
     for(int x=0; x<dimensions.x; x++){
-        std::vector<unsigned int> row = std::vector<unsigned int>();
+        std::vector<double> row = std::vector<double>();
         for(int y=0; y<dimensions.y; y++){
             row.push_back(height);
         }
