+++ src/terrain/terrain.cpp
@@ -5,17 +5,17 @@
 Terrain::Terrain(unsigned int x, unsigned int y)
     : size_x(x), size_y(y)
 {
-    heightmap = std::vector<std::vector<unsigned int>>(size_x);
+    heightmap = std::vector<std::vector<double>>(size_x);
     for(int i=0; i<size_x; i++)
     {
-        heightmap[i] = std::vector<unsigned int>(size_y);
+        heightmap[i] = std::vector<double>(size_y);
     }
 }
 
 Terrain::Terrain(Heightmap _heightmap):
 size_x(_heightmap.capacity()), size_y(_heightmap.at(0).capacity())
 {
-    heightmap = std::vector<std::vector<unsigned int>>(size_x);
+    heightmap = std::vector<std::vector<double>>(size_x);
     for(int i=0; i<size_x; i++)
     {
         heightmap[i] = _heightmap.at(i);
